create index.html and update remaining files

+++ src/main/resources/db/migration/V1__Create_languages_table.sql
@@ -1,5 +1,5 @@
 create table languages (
-    id bigint unsigned primary key auto_increment,
+    id INTEGER unsigned primary key auto_increment,
     welcomeMsg varchar(100) not null,
     code varchar (3) not null
 );
